% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRkov.R
\name{mRkov-package}
\alias{mRkov-package}
\alias{mRkov}
\title{Corncob package documentation.}
\description{
Corncob provides methods for estimating and plotting count data. Specifically, corncob is designed to account for the challenges of modeling sequencing data from microbial abundance studies.
}
\details{
mRkov is an R package that allows users to quickly parse textual data to be simulated using markov chain monte carlo methods.

The development version of the package will be maintained on \url{https://github.com/serrat839/mRkov}.
}
