% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{A test function!}
\usage{
setup(api_key, api_secret, token, token_secret)
}
\arguments{
\item{love}{Do you lvoe cats? Default to true}
}
\description{
This function allows you to link your twitter to the twitter API
If you don't have twitter api keys, go to https://developer.twitter.com/ and 
create an account. Then, go to create an app, fill out the form, and you should be able to access your very own keys and tokens.
}
\examples{
test_function()
}
\keyword{cats}
