% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trump_tweets.R
\docType{data}
\name{trump_tweets}
\alias{trump_tweets}
\title{Trump Tweet Data}
\format{An object of class \code{list} of length 3.}
\usage{
data(trump_tweets)
}
\description{
A sample outcome from tweet_gettr scraping @realdonaldtrump
}
\keyword{datasets}
