% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_gettr.R
\name{tweet_gettr}
\alias{tweet_gettr}
\title{A test function!}
\usage{
tweet_gettr(handle, output)
}
\arguments{
\item{love}{Do you lvoe cats? Default to true}
}
\description{
This function allows you to express your love of cats.
}
\examples{
test_function()
}
\keyword{cats}
