% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trumpdata.R
\docType{data}
\name{tweets}
\alias{tweets}
\title{Trump Tweet Data}
\format{An object of class \code{list} of length 3.}
\usage{
data(tweets)
}
\description{
A sample outcome from tweet_gettr
}
\keyword{datasets}
