% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sentence.R
\name{make_sentence}
\alias{make_sentence}
\title{Make Sentence}
\usage{
make_sentence(corpus, prompt = "", n = 1)
}
\arguments{
\item{corpus}{List. A list output from tweet_gettr.}

\item{prompt}{String. A string containing a word to start the markov chain.}

\item{n}{Integer. The length of the sequence of words we want to match when generating a sentence.}
}
\description{
This function allows you to create a sentence from a body of words loaded by this package
}
\examples{
make_sentence(trump_tweets, prompt = "I")
}
\keyword{sentence}
