% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wordcloud.R
\name{make_wordcloud}
\alias{make_wordcloud}
\title{TwittR Setup}
\usage{
make_wordcloud(information, stops = "")
}
\arguments{
\item{information}{List. The output from tweet_gettr or read_text_file}

\item{stops}{String or vector of strings. Words to remove from the word cloud}
}
\description{
This function takes in the output from tweet_gettr or read_text_file and generates a word cloud from the most frequent words
words can be removed by using the stops parameter
}
\examples{
make_wordcloud(trump_tweets)

}
\keyword{api}
\keyword{setup,}
