% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text_file.R
\name{read_text_file}
\alias{read_text_file}
\title{Tweet GettR}
\usage{
read_text_file(text_file, delimiter = "\\n", sentiments = NULL)
}
\arguments{
\item{text_file}{String. Path to the text file you want to read in}

\item{delimiter}{String. String that dictates what is a new line}

\item{sentiments}{Dataframe. Dataframe containing words and their sentiments. Columns MUST be "word" and "sentiment."}
}
\value{
Returns a properly-formatted list that you can run make_sentence on
}
\description{
This function lets you scrape the timeline of any twitter user. You MUST run the setup function before using this function in order for it to work.
}
\keyword{Scraping}
\keyword{Twitter,}
