% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_twitteR.R
\name{setup_twitteR}
\alias{setup_twitteR}
\title{TwittR Setup}
\usage{
setup_twitteR()
}
\description{
This function allows you to link your twitter to the twitter API
If you don't have twitter api keys, go to https://developer.twitter.com/ and
create an account. Then, go to create an app, fill out the form, and you should be able to access your very own keys and tokens.
}
\examples{
setup_twitteR()
}
\keyword{api}
\keyword{setup,}
