% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_gettr.R
\name{tweet_gettr}
\alias{tweet_gettr}
\title{Tweet GettR}
\usage{
tweet_gettr(
  handle,
  token = NULL,
  output = "",
  n = 3200,
  includeRts = FALSE,
  sentiments = NULL
)
}
\arguments{
\item{handle}{The handle of the twitter user's timeline you wish to scrape.}

\item{token}{Token. The token generated by setup_twitteR.}

\item{output}{The location of an output text file that you can re-use if you want a re-usable file of tweets. Defaults to a blank string, no output saved.}

\item{n}{Integer. Defaults to `3200`. The number of tweets to pull.}

\item{includeRts}{Boolean. Defaults to `FALSE`. Whether or not to include re-tweets.}

\item{sentiments}{Dataframe. Dataframe containing words and their sentiments. Columns MUST be "word" and "sentiment"}
}
\value{
Returns a properly-formatted list that you can run make_sentence on
}
\description{
This function lets you scrape the timeline of any twitter user. You MUST run the setup function before using this function in order for it to work.
}
\examples{
tweet_gettr("@realDonaldTrump", "./trump.txt")
}
\keyword{Scraping}
\keyword{Twitter,}
